# docs/swagger.yaml
basePath: /
definitions:
  dto.ErrorOutput:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  dto.FibonacciSpiralMatrixDto:
    properties:
      rows:
        items:
          items:
            type: integer
          type: array
        type: array
    type: object
  dto.HealthOutput:
    properties:
      details:
        additionalProperties: true
        type: object
      status:
        example: ok
        type: string
    type: object
  dto.MatrixInput:
    properties:
      column:
        type: string
      row:
        type: string
    required:
    - column
    - row
    type: object
  dto.SignupRequestDto:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - role
    type: object
  fibonacci-spiral-matrix-go_internal_api_handler.LoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  fibonacci-spiral-matrix-go_internal_api_handler.LoginResponse:
    properties:
      jwt:
        type: string
      roles:
        type: string
      userdetailsemail:
        type: string
      userdetailsid:
        type: integer
      userdetailsusername:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_api_handler.LoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  internal_api_handler.LoginResponse:
    properties:
      jwt:
        type: string
      roles:
        type: string
      userdetailsemail:
        type: string
      userdetailsid:
        type: integer
      userdetailsusername:
        type: string
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the fibonacci spiral matrix restful api server.
  title: Fibonacci Spiral Matrix API
  version: 1.0.0
paths:
  /api/status:
    get:
      consumes:
      - application/json
      description: Get api pulse status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthOutput'
      summary: Check api status
      tags:
      - Health Check
  /api/user/spiral:
    get:
      consumes:
      - application/json
      description: Get matrix by row and column
      parameters:
      - description: Get spiral matrix
        in: body
        name: matrix
        required: true
        schema:
          $ref: '#/definitions/dto.MatrixInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FibonacciSpiralMatrixDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      summary: Fetch a spiral matrix
      tags:
      - matrix
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login a user by username and password
      parameters:
      - description: Login user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/fibonacci-spiral-matrix-go_internal_api_handler.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fibonacci-spiral-matrix-go_internal_api_handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      summary: Login a user
      tags:
      - login
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: SignUp a user with field email, username, password
      parameters:
      - description: Insert user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SignupRequestDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorOutput'
      summary: SignUp a user
      tags:
      - Auth
swagger: "2.0"